Removed old tempToRGB method.
Colors are now either randomly selected for you or you can pass in an
hex/rgb value. Kinda want you to pass in a hex value, much easier to
deal with.

Something seems to be wrong with how i'm determining the intensity in
the HSL value. Sometimes, it comes back all black. hurp